Add files via upload
 Car breakdowns are one of the common problems that most of us face on frequent base.  Lack of time for car  periodic inspection and emergency cases as well.
Car Garage project will provides repair of any car malfunction and maintenance services with technicians equipped on the road (replacement of car batteries, car oils, filters of all kinds, tire replacement service, car wash, etc.).
Furthermore, the project will distinguish by reaching the customer as quickly as possible to the car in its location whether at home, workplace or on the side of the road.  Moreover, the target is the society in general at any time anywhere, given that people are exposed to these problems on a daily basis. 
At the end, the project depends on the foundations of good work according to the strategies of successful work represented in the principles of speed and mastery. Car Garage will also increase the job opportunities and provide easier life.